Move domain/contracts into clearer folders
