+++ src/ActorApi.Api/Domains/ClientSelection.cs
@@ -3,17 +3,6 @@
 namespace ActorApi.Api.Domains
 {
     /// <summary>
-    /// Data Request Model
-    /// </summary>
-    public class DataActorRequest
-    {
-        public required ClientSelection ClientSelection {  get; init; }
-        public string? Param1 { get; set; }
-        public string? Param2 { get; set; }
-        public string? Header1 { get; set; }
-        public string? Header2 { get; set; }
-    }
-    /// <summary>
     /// List of available Clients
     /// </summary>
     public enum ClientSelection
