+++ src/ActorApi.Api/Services/IActorService.cs
@@ -1,4 +1,4 @@
-﻿using ActorApi.Api.Domains;
+﻿using ActorApi.Api.Contracts;
 
 namespace ActorApi.Api.Services
 {
