+++ tests/ActorApi.Tests/Services/ActorServiceTests.cs
@@ -1,4 +1,5 @@
 ﻿using ActorApi.Api.Clients;
+using ActorApi.Api.Contracts;
 using ActorApi.Api.Domains;
 using ActorApi.Api.Services;
 using ActorApi.Api.Services.Resolvers;
